// src/types/Producto.tsx
export interface Producto {
    idproducto: number
    nombre: string
    precio: number
    preciorebajado: number
    imagenchica: string
    imagengrande: string
    categoria: string
    detalle: string
    unidadesenexistencia: number
    proveedor: string
    telefono: string
    pais: string
    promedioestrellas: number
    totalcalificaciones: number
    descripcion: string
}